fix: mdx tables + port gatsby core to ESM
