+++ packages/gatsby-theme-carbon/gatsby-config.mjs
@@ -1,22 +1,32 @@
-const path = require('path');
-const { gray100 } = require('@carbon/elements');
-const defaultLunrOptions = require('./config/lunr-options');
+import path from 'path';
+import { gray100 } from '@carbon/elements';
+import defaultLunrOptions from './config/lunr-options.js';
+/*
+This file has been ported to ES module just because of remark-gfm plugin
+More info - https://www.reddit.com/r/gatsbyjs/comments/107nnj2/gatsby_5_mdx2_support_for_tables/
+Also we have tied the version to be 3.0.1 as the latest version will uses markdown 3 is not fully supported by gatsby v5 at the moment
+More info - https://github.com/mdx-js/mdx/issues/2379#issuecomment-1933035305
+*/
+import remarkGfm from 'remark-gfm';
+import { fileURLToPath } from 'url';
+import { dirname } from 'path';
 
+const __dirname = dirname(fileURLToPath(import.meta.url));
 const carbonThemes = {
-  white: require.resolve('./src/styles/internal/white.scss'),
-  g10: require.resolve('./src/styles/internal/g10.scss'),
-  dark: require.resolve('./src/styles/internal/g100.scss'),
+  white: './src/styles/internal/white.scss',
+  g10: './src/styles/internal/g10.scss',
+  dark: './src/styles/internal/g100.scss',
 };
 
-module.exports = (themeOptions) => {
-  const repositoryDefault = {
-    baseUrl: '',
-    subDirectory: '',
-    branch: 'main',
-  };
+const repositoryDefault = {
+  baseUrl: '',
+  subDirectory: '',
+  branch: 'main',
+};
 
-  const defaultTheme = { homepage: 'dark', interior: 'g10' };
+const defaultTheme = { homepage: 'dark', interior: 'g10' };
 
+export default (themeOptions) => {
   const {
     theme: themeOption,
     isSearchEnabled = true,
@@ -36,7 +46,6 @@ module.exports = (themeOptions) => {
   } = themeOptions;
 
   const theme = { ...defaultTheme, ...themeOption };
-
   const optionalPlugins = [];
 
   if (mediumAccount) {
@@ -111,7 +120,7 @@ module.exports = (themeOptions) => {
             ...gatsbyRemarkPlugins,
           ],
           mdxOptions: {
-            remarkPlugins,
+            remarkPlugins: [...remarkPlugins, remarkGfm],
           },
           // defaultLayouts: {
           //   default: require.resolve('./src/templates/Default.js'),
@@ -131,8 +140,8 @@ module.exports = (themeOptions) => {
         resolve: `@garcia-enterprise/gatsby-plugin-sass-resources`,
         options: {
           resources: [
-            require.resolve('./src/styles/internal/resources.scss'),
-            carbonThemes[theme.interior],
+            path.resolve(__dirname, `./src/styles/internal/resources.scss`),
+            path.resolve(__dirname, carbonThemes[theme.interior]),
           ],
         },
       },
@@ -162,9 +171,10 @@ module.exports = (themeOptions) => {
           display: 'browser',
           icon: iconPath
             ? path.resolve(iconPath)
-            : require.resolve('./src/images/favicon.svg'),
+            : path.resolve('./src/images/favicon.svg'),
         },
       },
     ].concat(optionalPlugins),
   };
 };
+
