+++ packages/gatsby-theme-carbon/package.json
@@ -62,6 +62,7 @@
     "react-helmet": "^6.1.0",
     "react-transition-group": "^4.4.5",
     "react-use": "^17.5.0",
+    "remark-gfm": "3.0.1",
     "sass": "^1.70.0",
     "slugify": "^1.6.6",
     "smooth-scroll": "^16.1.3",
